Added reading/writing blocks of signed BCD

+++ src/fins_utils.c
@@ -102,7 +102,8 @@ static bool valid_dos_filename_char[] = {
  *
  * The function finslib_bcd_to_int() converts a BCD encode value to a binary
  * integer. If undefined BCD characters are present in the input, the value
- * INT32_MAX is returned instead of the converted value;
+ * INT16_MAX is returned instead of the converted value when a 16 bit value
+ * was requested and INT32_MAX for other data sizes.
  */
 
 int32_t finslib_bcd_to_int( uint32_t value, int type ) {
@@ -118,7 +119,7 @@ int32_t finslib_bcd_to_int( uint32_t value, int type ) {
 
 		case FINS_DATA_TYPE_BCD16 :
 
-			if ( value & 0xFFFF0000 ) return INT32_MAX;
+			if ( value & 0xFFFF0000 ) return INT16_MAX;
 
 			minval   = 0;
 			maxval   = 9999;
@@ -140,7 +141,7 @@ int32_t finslib_bcd_to_int( uint32_t value, int type ) {
 
 		case FINS_DATA_TYPE_SBCD16_0 :
 
-			if ( value & 0xFFFFE000 ) return INT32_MAX;
+			if ( value & 0xFFFFE000 ) return INT16_MAX;
 
 			minval   = -999;
 			maxval   =  999;
@@ -166,7 +167,7 @@ int32_t finslib_bcd_to_int( uint32_t value, int type ) {
 
 		case FINS_DATA_TYPE_SBCD16_1 :
 
-			if ( value & 0xFFFF0000 ) return INT32_MAX;
+			if ( value & 0xFFFF0000 ) return INT16_MAX;
 
 			minval   = -7999;
 			maxval   = 7999;
@@ -190,7 +191,7 @@ int32_t finslib_bcd_to_int( uint32_t value, int type ) {
 
 		case FINS_DATA_TYPE_SBCD16_2 :
 
-			if ( value & 0xFFFF0000 ) return INT32_MAX;
+			if ( value & 0xFFFF0000 ) return INT16_MAX;
 
 			minval   = -999;
 			maxval   = 9999;
@@ -259,7 +260,7 @@ int32_t finslib_bcd_to_int( uint32_t value, int type ) {
 				case 0x00009000 : negative = false; value =  value & 0x0000FFFF;               break;
 				case 0x0000A000 : negative = true;  value = (value & 0x00000FFF) | 0x00001000; break;
 				case 0x0000F000 : negative = true;  value =  value & 0x00000FFF;               break;
-				default         : return INT32_MAX;
+				default         : return INT16_MAX;
 			}
 			
 			break;
@@ -322,8 +323,8 @@ int32_t finslib_bcd_to_int( uint32_t value, int type ) {
  *
  * The function finslib_int_to_bcd() converts an binary coded integer to a BCD
  * value. If the function succeeds, the converted value is returned. Otherwise
- * the returned value is UINT32_MAX to indicate that the value can not be
- * encoded in BCD.
+ * the returned value is INT16_MAX or INT32_MAX to indicate that the value can
+ * not be encoded in BCD.
  */
 
 uint32_t finslib_int_to_bcd( int32_t value, int type ) {
@@ -339,7 +340,7 @@ uint32_t finslib_int_to_bcd( int32_t value, int type ) {
 
 		case FINS_DATA_TYPE_BCD16 :
 
-			if ( value < 0  ||  value > 9999 ) return UINT32_MAX;
+			if ( value < 0  ||  value > 9999 ) return INT16_MAX;
 
 			break;
 
@@ -347,7 +348,7 @@ uint32_t finslib_int_to_bcd( int32_t value, int type ) {
 
 		case FINS_DATA_TYPE_BCD32 :
 
-			if ( value < 0  ||  value > 99999999 ) return UINT32_MAX;
+			if ( value < 0  ||  value > 99999999 ) return INT32_MAX;
 
 			break;
 
@@ -355,7 +356,7 @@ uint32_t finslib_int_to_bcd( int32_t value, int type ) {
 
 		case FINS_DATA_TYPE_SBCD16_0 :
 
-			if ( value < -999  ||  value > 999 ) return UINT32_MAX;
+			if ( value < -999  ||  value > 999 ) return INT16_MAX;
 
 			if ( value < 0 ) { neg_mask = 0x00001000; value = -value; }
 
@@ -365,7 +366,7 @@ uint32_t finslib_int_to_bcd( int32_t value, int type ) {
 
 		case FINS_DATA_TYPE_SBCD32_0 :
 
-			if ( value < -9999999  ||  value > 9999999 ) return UINT32_MAX;
+			if ( value < -9999999  ||  value > 9999999 ) return INT32_MAX;
 
 			if ( value < 0 ) { neg_mask = 0x10000000; value = -value; }
 
@@ -375,7 +376,7 @@ uint32_t finslib_int_to_bcd( int32_t value, int type ) {
 
 		case FINS_DATA_TYPE_SBCD16_1 :
 
-			if ( value < -7999  ||  value > 7999 ) return UINT32_MAX;
+			if ( value < -7999  ||  value > 7999 ) return INT16_MAX;
 
 			if ( value < 0 ) { neg_mask = 0x00008000; value = -value; }
 
@@ -385,7 +386,7 @@ uint32_t finslib_int_to_bcd( int32_t value, int type ) {
 
 		case FINS_DATA_TYPE_SBCD32_1 :
 
-			if ( value < -79999999  ||  value > 79999999 ) return UINT32_MAX;
+			if ( value < -79999999  ||  value > 79999999 ) return INT32_MAX;
 
 			if ( value < 0 ) { neg_mask = 0x80000000; value = -value; }
 
@@ -395,7 +396,7 @@ uint32_t finslib_int_to_bcd( int32_t value, int type ) {
 
 		case FINS_DATA_TYPE_SBCD16_2 :
 
-			if ( value < -999   ||  value > 9999 ) return UINT32_MAX;
+			if ( value < -999   ||  value > 9999 ) return INT16_MAX;
 
 			if ( value < 0 ) { neg_mask = 0x0000F000; value = -value; }
 
@@ -405,7 +406,7 @@ uint32_t finslib_int_to_bcd( int32_t value, int type ) {
 
 		case FINS_DATA_TYPE_SBCD32_2 :
 
-			if ( value < -9999999  ||  value > 99999999 ) return UINT32_MAX;
+			if ( value < -9999999  ||  value > 99999999 ) return INT32_MAX;
 
 			if ( value < 0 ) { neg_mask = 0xF0000000; value = -value; }
 
@@ -415,7 +416,7 @@ uint32_t finslib_int_to_bcd( int32_t value, int type ) {
 
 		case FINS_DATA_TYPE_SBCD16_3 :
 
-			if ( value < -1999  ||  value > 9999 ) return UINT32_MAX;
+			if ( value < -1999  ||  value > 9999 ) return INT16_MAX;
 
 			if      ( value < -999 ) { neg_mask = 0x0000A000; value = -value; }
 			else if ( value <    0 ) { neg_mask = 0x0000F000; value = -value; }
@@ -426,7 +427,7 @@ uint32_t finslib_int_to_bcd( int32_t value, int type ) {
 
 		case FINS_DATA_TYPE_SBCD32_3 :
 
-			if ( value < -19999999  ||  value > 99999999 ) return UINT32_MAX;
+			if ( value < -19999999  ||  value > 99999999 ) return INT32_MAX;
 
 			if      ( value < -9999999 ) { neg_mask = 0xA0000000; value = -value; }
 			else if ( value <        0 ) { neg_mask = 0xF0000000; value = -value; }
@@ -439,7 +440,7 @@ uint32_t finslib_int_to_bcd( int32_t value, int type ) {
 
 	while ( value != 0 ) {
 
-		if ( shiftval > 24 ) return UINT32_MAX;
+		if ( shiftval > 24 ) return INT32_MAX;
 
 		addval    = inttobcd_lut[ value % 100 ];
 		retval   += addval << shiftval;
